+++ test/ArborX_BoostGeometryAdapters.hpp
@@ -33,7 +33,7 @@ struct tag<ArborX::Point>
 template <>
 struct coordinate_type<ArborX::Point>
 {
-  typedef double type;
+  typedef float type;
 };
 
 template <>
@@ -50,9 +50,9 @@ struct dimension<ArborX::Point> : boost::mpl::int_<3>
 template <size_t D>
 struct access<ArborX::Point, D>
 {
-  static inline double get(ArborX::Point const &p) { return p[D]; }
+  static inline float get(ArborX::Point const &p) { return p[D]; }
 
-  static inline void set(ArborX::Point &p, double value) { p[D] = value; }
+  static inline void set(ArborX::Point &p, float value) { p[D] = value; }
 };
 
 // Adapt ArborX::Box to Boost.Geometry
@@ -71,9 +71,9 @@ struct point_type<ArborX::Box>
 template <size_t D>
 struct indexed_access<ArborX::Box, min_corner, D>
 {
-  static inline double get(ArborX::Box const &b) { return b.minCorner()[D]; }
+  static inline float get(ArborX::Box const &b) { return b.minCorner()[D]; }
 
-  static inline void set(ArborX::Box &b, double value)
+  static inline void set(ArborX::Box &b, float value)
   {
     b.minCorner()[D] = value;
   }
@@ -82,9 +82,9 @@ struct indexed_access<ArborX::Box, min_corner, D>
 template <size_t D>
 struct indexed_access<ArborX::Box, max_corner, D>
 {
-  static inline double get(ArborX::Box const &b) { return b.maxCorner()[D]; }
+  static inline float get(ArborX::Box const &b) { return b.maxCorner()[D]; }
 
-  static inline void set(ArborX::Box &b, double value)
+  static inline void set(ArborX::Box &b, float value)
   {
     b.maxCorner()[D] = value;
   }
